requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-WTF
python-dotenv
cryptography
email-validator
bleach
eventlet
gunicorn